requirements.txt
black
Django~=3.0
django-crispy-forms~=1.9
isort<5,>=4.2.5
pylint
pylint-django
rope
whitenoise~=5.1
Pillow~=7.2